pyproject.toml
[project]
name = "bash-landing"
version = "1.0.0"
description = "Terminal-based lunar lander game with ANSI half-block pixel rendering"
requires-python = ">=3.10"
dependencies = [
    "pyaudio",
]

[project.optional-dependencies]
image = ["Pillow"]
record = ["asciinema"]
pygame = ["pygame"]

[project.scripts]
bash-landing = "game.main:main"
